% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSequence-package.R
\docType{package}
\name{dbSequence-package}
\alias{dbSequence-package}
\title{dbSequence: DuckDB-Backed Interface for Large-Scale Genomic Data}
\description{
dbSequence provides DuckDB-backed infrastructure for genomic interval and
sequence-derived data that are too large to work with comfortably in memory.
It imports common genomics file formats into DuckDB, keeps range operations
lazy, and interoperates with Bioconductor classes such as
\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbverse-org/dbSequence}
  \item Report bugs at \url{https://github.com/dbverse-org/dbSequence/issues}
}

}
\author{
\strong{Maintainer}: Edward C. Ruiz \email{ecr7407@gmail.com} (\href{https://orcid.org/0000-0002-9174-5387}{ORCID})

Other contributors:
\itemize{
  \item Ruben Dries [reviewer]
  \item National Human Genome Research Institute (1R01HG014487-01) [funder]
}

}
\keyword{internal}
