% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{file_source}
\alias{file_source}
\title{Get file source from dbSequence (deprecated)}
\usage{
file_source(x)
}
\arguments{
\item{x}{(required) A dbSequence object}
}
\value{
character: the source file path
}
\description{
Use fileSource() instead. This function is deprecated.
}
