% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/validation.R
\name{fileSource}
\alias{fileSource}
\alias{fileSource,dbSequence-method}
\alias{fileSource<-,dbSequence-method}
\title{Get file source from dbSequence}
\usage{
fileSource(object)

fileSource(object)

\S4method{fileSource}{dbSequence}(object)

\S4method{fileSource}{dbSequence}(object) <- value
}
\arguments{
\item{object}{A dbSequence object}

\item{x}{(required) A dbSequence object}
}
\value{
character: the source file path

The file_source value
}
\description{
Extract the file source path from a dbSequence object.
}
