% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,BamFile,missing,ANY-method}
\alias{import,BamFile,missing,ANY-method}
\title{Import BAM files into DuckDB}
\usage{
\S4method{import}{BamFile,missing,ANY}(con, format, text, dest, table_name = "bam_data", .conn = NULL, ...)
}
\description{
Import BAM files into DuckDB
}
