% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,BEDFile,missing,ANY-method}
\alias{import,BEDFile,missing,ANY-method}
\title{Import BED files into DuckDB}
\usage{
\S4method{import}{BEDFile,missing,ANY}(con, format, text, dest, table_name = "bed_data", ...)
}
\arguments{
\item{con}{A file connection object.}

\item{format}{Import format; unused for these methods.}

\item{text}{Text input; unused for these methods.}

\item{dest}{A \code{DuckDBFile} destination.}

\item{table_name}{Table name to create.}

\item{...}{Additional arguments.}
}
\value{
A dbSequence object backed by a DuckDB table.
}
\description{
Import BED files into DuckDB
}
