% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,FastaFile,missing,ANY-method}
\alias{import,FastaFile,missing,ANY-method}
\title{Import FASTA files into DuckDB}
\usage{
\S4method{import}{FastaFile,missing,ANY}(con, format, text, dest, table_name = "fasta_data", ...)
}
\description{
Import FASTA files into DuckDB
}
