% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,GFFFile,missing,ANY-method}
\alias{import,GFFFile,missing,ANY-method}
\title{Import GFF files into DuckDB}
\usage{
\S4method{import}{GFFFile,missing,ANY}(con, format, text, dest, table_name = "gff_data", ...)
}
\description{
Import GFF files into DuckDB
}
