% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,GTFFile,missing,ANY-method}
\alias{import,GTFFile,missing,ANY-method}
\title{Import GTF files into DuckDB}
\usage{
\S4method{import}{GTFFile,missing,ANY}(con, format, text, dest, table_name = "gtf_data", ...)
}
\description{
Import GTF files into DuckDB
}
