% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show,dbSequence-method}
\alias{show,dbSequence-method}
\title{Show method for dbSequence objects}
\usage{
\S4method{show}{dbSequence}(object)
}
\arguments{
\item{object}{A dbSequence object}
}
\value{
Invisibly returns NULL.
}
\description{
Display a summary of the dbSequence object and preview the table data
}
