% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{table_name}
\alias{table_name}
\title{Get table name from dbSequence (deprecated)}
\usage{
table_name(x)
}
\arguments{
\item{x}{(required) A dbSequence object}
}
\value{
character: the table name
}
\description{
Use tableName() instead. This function is deprecated.
}
