% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tableName}
\alias{tableName}
\title{Get table name from dbSequence}
\usage{
tableName(x)
}
\arguments{
\item{x}{(required) A dbSequence object}
}
\value{
character: the table name
}
\description{
Extract the table name from a dbSequence object.
}
