% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyranges-compat.R
\name{as_granges}
\alias{as_granges}
\alias{as_granges.default}
\alias{as_granges.dbSequence}
\title{Convert objects to GRanges}
\usage{
as_granges(.data, ..., keep_mcols = TRUE)

\method{as_granges}{default}(.data, ..., keep_mcols = TRUE)

\method{as_granges}{dbSequence}(.data, ..., keep_mcols = TRUE)
}
\arguments{
\item{.data}{A dbSequence object}

\item{...}{Additional arguments (currently unused)}

\item{keep_mcols}{logical: whether to keep metadata columns (default: TRUE)}
}
\value{
A GRanges object

A GRanges object
}
\description{
S3 generic for converting various objects to GRanges.
This generic is provided for plyranges compatibility when plyranges
is not loaded.

S3 method for plyranges compatibility. Converts a dbSequence
object to a GRanges object by collecting data from DuckDB.
}
\details{
This method enables plyranges-style workflows:
\code{db_seq \%>\% as_granges()}

Note that this collects all data from DuckDB into memory. For large
datasets, consider filtering first using dplyr verbs.
}
\examples{
# Import BED file to dbSequence
bed <- system.file("extdata", "example.bed", package = "dbSequence")
db_seq <- read_bed(bed)

# Convert to GRanges (collects data)
gr <- as_granges(db_seq)

}
