% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{asRanges}
\alias{asRanges}
\title{Extract genomic ranges from dbSequence objects}
\usage{
asRanges(x, ...)
}
\arguments{
\item{x}{(required) A dbSequence object}

\item{...}{(optional) Additional arguments}
}
\value{
A view of the dbSequence with only range columns (seqnames, start, end, strand)
}
\description{
Extract genomic ranges from dbSequence objects
}
\examples{
bed <- system.file("extdata", "example.bed", package = "dbSequence")
db_path <- tempfile(fileext = ".duckdb")
db_seq <- read_bed(bed, dest = DuckDBFile(db_path), lazy = FALSE)
asRanges(db_seq)

}
