% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{dbSequence-class}
\alias{dbSequence-class}
\title{dbSequence objects}
\description{
S4 class for genomic sequences stored in DuckDB

Extends dbData to wrap genomic data stored in DuckDB tables.
}
\section{Slots}{

\describe{
\item{\code{value}}{dplyr tbl representing the genomic data in the database (inherited from dbData)}

\item{\code{name}}{character table name in database (inherited from dbData)}

\item{\code{file_source}}{character source file path or identifier (immutable after creation)}
}}

