% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{fileSource<-}
\alias{fileSource<-}
\title{Replacement method for file_source - BLOCKED}
\usage{
fileSource(object) <- value
}
\arguments{
\item{object}{A dbSequence object}

\item{value}{New value (will be rejected)}
}
\description{
This method prevents modification of the file_source slot
after object creation. The file_source is set during import and should
not be changed to maintain data integrity.
}
