% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{fileSource}
\alias{fileSource}
\alias{fileSource,dbSequence-method}
\alias{fileSource<-,dbSequence-method}
\title{Accessor for file_source slot}
\usage{
fileSource(object)

\S4method{fileSource}{dbSequence}(object)

\S4method{fileSource}{dbSequence}(object) <- value
}
\arguments{
\item{object}{A dbSequence object}

\item{value}{New value (will be rejected)}
}
\value{
The file_source value
}
\description{
Accessor for file_source slot
}
\examples{
bed <- system.file("extdata", "example.bed", package = "dbSequence")
db_seq <- read_bed(bed)
fileSource(db_seq)
try(fileSource(db_seq) <- "other.bed")

}
