% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,BEDFile,missing,ANY-method}
\alias{import,BEDFile,missing,ANY-method}
\title{Import BED files into DuckDB}
\usage{
\S4method{import}{BEDFile,missing,ANY}(con, format, text, dest, table_name = "bed_data", ...)
}
\description{
Import BED files into DuckDB
}
