% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,character,missing,ANY-method}
\alias{import,character,missing,ANY-method}
\title{Import character file paths into DuckDB (auto-detect format)}
\usage{
\S4method{import}{character,missing,ANY}(con, format, text, dest, table_name = NULL, ...)
}
\description{
Import character file paths into DuckDB (auto-detect format)
}
