% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import,VcfFile,missing,ANY-method}
\alias{import,VcfFile,missing,ANY-method}
\title{Import VCF files into DuckDB}
\usage{
\S4method{import}{VcfFile,missing,ANY}(con, format, text, dest, table_name = "vcf_data", ...)
}
\description{
Import VCF files into DuckDB
}
