% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show,DuckDBFile-method}
\alias{show,DuckDBFile-method}
\title{Show method for DuckDBFile objects}
\usage{
\S4method{show}{DuckDBFile}(object)
}
\arguments{
\item{object}{A DuckDBFile object}
}
\value{
Invisibly returns NULL.
}
\description{
Display information about the DuckDBFile object.
}
